+++ src/main/java/io/redstudioragnarok/valkyrie/asm/ValkyriePlugin.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Map;
 
 @IFMLLoadingPlugin.TransformerExclusions("io.redstudioragnarok.valkyrie.asm")
-public class FMLPlugin implements IFMLLoadingPlugin, IEarlyMixinLoader {
+public class ValkyriePlugin implements IFMLLoadingPlugin, IEarlyMixinLoader {
 
 	@Override
 	public String[] getASMTransformerClass() {