+++ src/main/java/io/redstudioragnarok/valkyrie/mixininterfaces/IArrayDequeRenderChunks.java
@@ -1,0 +1,10 @@
+package io.redstudioragnarok.valkyrie.mixininterfaces;
+
+import net.minecraft.client.renderer.chunk.RenderChunk;
+
+import java.util.ArrayDeque;
+
+public interface IArrayDequeRenderChunks {
+
+    ArrayDeque<RenderChunk> valkyrie$renderChunks = new ArrayDeque<>();
+}