+++ Changelog.md
@@ -12,10 +12,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added warning for using OptiFine
 - Built in support for Mantle for the MC-67532 fix
 - When in a deobfuscated environment, "Development Environment" will be added after the name of the window
+- Added terrain wireframe debug config
 
 ### Changed
 
 - Changed the upper limit for the `Zoom Multiplier` to 10
+- Reduced default cloud render distance
 
 ### Fixed
 
@@ -27,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Optimized `RenderGlobal#setupTerrain` which improves rendering speed
 - Optimized `RenderGlobal#getRenderChunkOffset` which improves rendering speed
+- Optimized `ChunkRenderContainer#preRenderChunk` which improves rendering speed
 - Optimized `ViewFrustum` which reduce FPS drops when blocks are updated, improves the speed of loading render chunks and reduce lag when loading renderer (Changing graphics settings, loading into world)
 
 ## Valkyrie Version 0.1.3 Changelog 2023-06-16